Refactors all i18n related core app functionality into kolibri.utils.i18n. Adds createTranslator function to create a translator object for use in JS files.
