+++ kolibri/core/assets/src/utils/import-intl-locale.js
